trying to connect all the pieces
