+++ app/lpc55xpresso/app.toml
@@ -146,6 +146,34 @@ task-slots = ["jefe"]
 stacksize = 1200
 extern-regions = ["sram2"]
 
+[tasks.sprot]
+name = "drv-lpc55-sprot-server"
+priority = 6
+max-sizes = {flash = 32768, ram = 32768}
+uses = ["flexcomm8", "bootrom"]
+features = ["spi0"]
+start = true
+notifications = ["spi-irq"]
+interrupts = {"flexcomm8.hs_spi" = "spi-irq"}
+stacksize = 16384
+task-slots = ["gpio_driver", "syscon_driver", "update_server"]
+
+[tasks.sprot.config]
+pins = [
+    # HS_SPI_MOSI = P0_26 = FUN9
+    { pin = { port = 0, pin = 26}, alt = 9},
+    # HS_SPI_MISO = P1_3 = FUN6
+    { pin = { port = 1, pin = 3}, alt = 6},
+    # HS_SPI_SCK = P1_2 = FUN6
+    { pin = { port = 1, pin = 2}, alt = 6},
+    # HS_SSEL1 = P1_1 = FUN5
+    { pin = { port = 1, pin = 1}, alt = 5},
+    # ROT_IRQ = P0_16 = FUN0
+    { name = "ROT_IRQ", pin = { port = 0, pin = 16}, alt = 0, direction = "output"},
+    # SP_RESET = P1_5 = FUN0 (not connected)
+    { name = "SP_RESET", pin = { port = 1, pin = 5}, alt = 0, direction = "input"},
+]
+
 [signing]
 enable-secure-boot = true
 enable-dice = true
